### 2019-01-28 #1214 @whedon
[PRE REVIEW]: FSharpGephiStreamer: An idiomatic bridge between F# and network visualization
**Submitting author:** @muehlhaus (<a href="http://orcid.org/0000-0003-3925-6778">Timo Mühlhaus</a>)
**Repository:** <a href="https://github.com/CSBiology/FSharpGephiStreamer" target ="_blank">https://github.com/CSBiology/FSharpGephiStreamer</a>
**Version:** v1.1.0
