### 2019-01-30 #1214 @arfon
This looks to be a bug in our LaTeX template at the `#` in the title. I've been trying to fix this for the last hour and haven't found an easy fix yet sorry.
