### 2019-02-06 #1214 @muehlhaus
The article proof looks fine. Thank you for fixing the pdf generator that it can cope with our article.
