### 2019-02-09 #1214 @yochannah
Hi @tpetricek - I know you're busy reviewing another item on JOSS at the minute so I won't ask you to look at another (unless you want to!) - but I was wondering if you could recommend any F#-ers who might be interested in reviewing this?
