### 2019-03-11 #1214 @kMutagene
Hi @yochannah,
Thank you for your continued efforts in finding a reviewer for our paper. I guess a factor why it is so hard finding people for this may be that all the fsharpers go straight to industry, which in my opinion is a pity as we see F# as a language suited very well for research in bioinformatics (and for all kinds of data analysis).
