### 2019-05-12 #1214 @labarba
Hi folks! @yochannah will be stepping down from the JOSS board, due to other obligations making it hard for her to volunteer time currently. I would like to reassign this submission to another editor.

@danielskatz — given current loads, could you take over on this one?
