### 2019-05-12 #1214 @danielskatz
👋 @TahaHachana, @dsyme, @jackfoxy, @tpetricek - would one of you be willing be the 2nd reviewer for this JOSS submission?  (See our [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) if needed)
