### 2019-05-13 #1214 @danielskatz
@PrashantVaidyanathan - Thanks! You will see as we start the review that a JOSS review is not a one-time read, but a set of suggestions and requests to the author, and an iterative process with the author to improve the submission (both the paper and the software)
