### 2019-05-18 #1445 @danielskatz
👋 @cgravill & @PrashantVaidyanathan - Is there anything I can do to help with your reviews?

If you can at least check your "Conflict of interest" and "Code of Conduct" boxes when you feel ready, that would help me know that the processes are working and that you are starting :)

(And @PrashantVaidyanathan, yes, I do remember that you will not be working on this 'till next week due to travel)
