### 2019-05-18 #1445 @cgravill
Sure, completed those steps @danielskatz

How precisely should these requirements be taken? For example, they have a plain text license but it includes a .txt extension. Is that OK?
