### 2019-05-18 #1445 @danielskatz
You could ask that this be changed to have no extension, since the JOSS criteria says "Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?" and since the GitHub guidance for this is for it to either have no extension or be .md.  You could open an issue in the software repo to request this, and put in a comment here pointing to that issue.
