### 2019-05-18 CSBiology/FSharpGephiStreamer#8 @cgravill
REVIEW: license file naming
I recommend renaming the license file to not include an extension for compatibility with the GitHub and JOSS naming recommendations:

https://help.github.com/en/articles/adding-a-license-to-a-repository
https://joss.readthedocs.io/en/latest/review_criteria.html#software-license

https://github.com/openjournals/joss-reviews/issues/1445
