### 2019-05-18 #1445 @cgravill
I'm not aware it's a conflict of interest but for full transparency:
- @PrashantVaidyanathan and I are both employed in a group at the same institution: Microsoft Research
- I'm also a contributor to https://github.com/Microsoft/automatic-graph-layout an alternative network library to Gephi, and tools to connect to it from F#
