### 2019-05-18 #1445 @danielskatz
> @PrashantVaidyanathan and I are both employed in a group at the same institution: Microsoft Research

Are you both employed in the same group? Do you work together on the same projects?

> I'm also a contributor to https://github.com/Microsoft/automatic-graph-layout an alternative network library to Gephi, and tools to connect to it from F#

Expertise in the subject is good.

With both factors in mind, do you feel that you can provide a fair and unbiased review and advise?
