### 2019-05-18 #1445 @cgravill
Yes, we're both employed in this group https://www.microsoft.com/en-us/research/group/biological-computation/ and have overlap in projects.

We've not discussed the contents of this work.

I feel I'm able to give a fair and unbiased review.
