### 2019-05-18 CSBiology/FSharpGephiStreamer#9 @cgravill
REVIEW: installation instructions
It would be nice to include instructions for OpenJDK to ensure flexibility of usage under a permissive license. I was was able to run Gephi on Windows successfully with binaries installed from https://adoptopenjdk.net/

The only complication was Gephi didn't automatically detect the installation which meant I had to alter the gephi.conf (found at C:\Program Files\Gephi-0.9.2\etc\gephi.conf) to contain:

```
jdkhome="C:\Program Files\AdoptOpenJDK\jre-11.0.3.7-hotspot"
```

https://joss.readthedocs.io/en/latest/review_criteria.html#what-about-submissions-that-rely-upon-proprietary-languages-development-environments

The install instructions seem to assume a working F# installation. It might help newcomers to point towards https://fsharp.org/ for instructions on installing F# for their environment.

openjournals/joss-reviews#1445
