### 2019-05-19 #1445 @danielskatz
Thanks for providing this info, @cgravill

Given the particular circumstances of this software and this conflict, we (the associate editors-in-chief) have decided that we will note it and waive it for this review.  So please continue the review.
