### 2019-05-19 CSBiology/FSharpGephiStreamer#10 @kMutagene
Hi @cgravill ,

first of all thank you and @PrashantVaidyanathan for taking the time to review this submission.

You are correct, the first line should be skipped because those are the headers. I was able to reproduce the error and it is located in parsing the edge list. I will fix that asap.
