### 2019-05-19 CSBiology/FSharpGephiStreamer#10 @kMutagene
Should be fixed by #11
