### 2019-05-19 CSBiology/FSharpGephiStreamer#8 @kMutagene
Good catch, seems like this applies to all our projects since [ProjectScaffold](https://github.com/fsprojects/ProjectScaffold) (the project layout template we use for all our projects) initializes the license file with a .txt extension.

fixed with #12
