### 2019-05-27 CSBiology/FSharpGephiStreamer#9 @kMutagene
Hi @cgravill,
I have tested the usage of OpenJDK and it works for me the same way it does for you. The only difference is that i used the OpenJDK 8 LTS version as gephi specifies it wants Java 8 or 9 (i'm not sure if version numbers are equal here, as it seems to work for you aswell when using Version 11.). Ill add this option to the installation instructions.
