### 2019-06-01 CSBiology/FSharpGephiStreamer#15 @cgravill
Review: missing dependency on Newtonsoft.Json
**Describe the bug**
It looks like the NuGet has a missing dependency on Newtonsoft.Json

```
CreateJsonFailed
  "Could not load file or assembly 'Newtonsoft.Json, Version=11.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed'. The system cannot find the file specified."
```

I'm able to workaround this which I did by installing Newtonsoft.Json but you should express the dependency on the NuGet

**OS and framework information (please complete the following information):**
 - OS: Windows 10
 - .Net core SDK version: 2.2.203

https://github.com/openjournals/joss-reviews/issues/1445
