### 2019-06-01 CSBiology/FSharpGephiStreamer#16 @cgravill
Review: I see no nodes / Http_CreateRequestFailed
**Describe the bug**
When attempting to create nodes I'm getting this exception: Http_CreateRequestFailed

Note while I started with the install instructions as they are at: https://csbiology.github.io/FSharpGephiStreamer/exampleAnalysis.html I made some adjustments. What you're currently doing:

```fsharp
let addOboNode (node:GONode) =

    let nodeConverter (node:GONode) =
        [
            Grammar.Attribute.Label (sprintf "%s | %s {%s}" node.Id node.TermDescription node.NameSpace);
            Grammar.Attribute.Size  10.;
            Grammar.Attribute.Color node.Color;
            Grammar.Attribute.UserDef ("UserData",node.TermDescription);
        ]
    Streamer.addNode nodeConverter node.Id node

goNodes |> List.map addOboNode
```

means that you're masking errors with the implicit ignore on the `List.map` (which generates warnings). I suggest you need to choose, if you're going to use the `Either` monad then it's important to include that in your examples

As to the actual error, I do have Gephi running but perhaps there's an access issue? Any guidance on debugging this?

**Expected behavior**
Nodes and edges into Gephi

**OS and framework information (please complete the following information):**
 - OS: Windows 10
 - .Net core SDK version [e.g. 2.2.203 ]

**Additional context**
This is possibly related to #15 perhaps I have the wrong version for you?

https://github.com/openjournals/joss-reviews/issues/1445
