### 2019-06-01 CSBiology/FSharpGephiStreamer#17 @cgravill
Review: example instructions
In the example instructions (https://csbiology.github.io/FSharpGephiStreamer/exampleAnalysis.html) you provide two approaches for loading nodes and edges. It's great that there's a locally reproducible version but also great to see your first alpha release of BioFSharp.

When taking the CSV route it gets a bit confusing. You have to get code later in the tutorial for e.g. createGONode but must only take part of the code from that section.

I suggest splitting and moving this section:

```fsharp

open FSharpGephiStreamer
open FSharpGephiStreamer.Colors

/// Simplified GO Term as node
type GONode = {
    //GO term (e.g. "GO:0000001")
    Id              : string
    //full term description e.g. "RNA polymerase I transcription factor complex"
    TermDescription : string
    //e.g. "biological process"
    NameSpace       : string
    //The color for the node
    Color           : Colors.Color
    }

/// Creates GONode
let createGONode id descr nameSpace col =
    {Id = id; TermDescription = descr; NameSpace = nameSpace ; Color = col}

/// Represents the Is_A relationship of GO terms as a directed edge
type GOEdge = {
    Id          : int
    Source      : string
    Target      : string
    TargetColor : Colors.Color
}

/// Creates GOEdge
let createGOEdge i source target tc = {Id = i; Source = source; Target = target; TargetColor = tc}
```

up before the CSV loading section then give people a choice to either load via CSV or using the Obo parser.

Hope this is clear and helps.

https://github.com/openjournals/joss-reviews/issues/1445
