### 2019-06-02 CSBiology/FSharpGephiStreamer#15 @kMutagene
hi @cgravill,

Newtonsoft.Json is listed in the dependencies file:
```
framework: auto-detect

source https://nuget.org/api/v2

nuget FSharp.Core
nuget Newtonsoft.Json //<- here
nuget Argu
...
...
```

Can i ask you to check 2 things? There should be a nuget package downloaded to `src/packages`, whats the version of that file? Second, when you go to `C:\Users\YourUserName\.nuget\packages`, whats the version of the package there?
