### 2019-06-02 CSBiology/FSharpGephiStreamer#15 @cgravill
Ah I wasn't specific, I'm making use of NuGet you're building:
https://www.nuget.org/packages/FSharpGephiStreamer/

![image](https://user-images.githubusercontent.com/1099725/58764996-17904500-8566-11e9-93da-0fe43eee4400.png)

In C:\Users\<MyUserName>\.nuget\packages\Newtonsoft.Json\ I have many versions of Newtonsoft.Json from other projects.
