### 2019-06-02 CSBiology/FSharpGephiStreamer#15 @cgravill
Btw what I suspect is causing the issue is this file:
https://github.com/CSBiology/FSharpGephiStreamer/blob/2c5a3060061b8a33a98a329e193edab629216993/src/FSharpGephiStreamer/paket.template

where you're using a project type of "file" (https://fsprojects.github.io/Paket/template-files.html) so you have to be explicit about everything. I also use that in places, but "project" might work more easily in this instance YMMV.
