### 2019-06-03 CSBiology/FSharpGephiStreamer#15 @kMutagene
If i understand you correctly, the problem shows itself here right?

![image](https://user-images.githubusercontent.com/21338071/58786415-cb85e480-85e7-11e9-9f86-5367fd57d0a0.png)

so either an explicit dependency section like this is needed,

![image](https://user-images.githubusercontent.com/21338071/58786602-3c2d0100-85e8-11e9-91c5-3bb7bd813aa1.png)

or we change the type to project (which i like better). Expect an updated package soon.
