### 2019-06-03 CSBiology/FSharpGephiStreamer#16 @kMutagene
The streamer sends the edge and node data to a local server created by the graph streaming plugin.

So additionally to #15 (the creation of the Json failing due to the missing Newtonsoft.Json dependency), i could imagine two problems that could cause this:

 1. the streaming master server is not enabled :

![image](https://user-images.githubusercontent.com/21338071/58791706-fcb7e200-85f2-11e9-86a3-9022c51eb757.png)

 2. We stream nodes and edges to the following workspace per default:
   `"http://localhost:8080/workspace1?operation=updateGraph"` (workspace 1).
    if you have workspaces with other names you can change the target url by using the `setEnvironment` function as seen here :

![image](https://user-images.githubusercontent.com/21338071/58791873-528c8a00-85f3-11e9-9ec2-7e48879bd940.png)

Edit: I just saw that the information that you have to create a workspace in gephi is missing from the docs, ill fix that.
