### 2019-06-03 CSBiology/FSharpGephiStreamer#17 @kMutagene
Good point, I can see why this may be confusing. Updated docs with #18
