### 2019-06-03 CSBiology/FSharpGephiStreamer#15 @cgravill
Yes, precisely, that shows a lack of dependency on the existing NuGet.
