### 2019-06-03 a3a32b5
Bump version to 1.1.1
