### 2019-06-03 CSBiology/FSharpGephiStreamer#15 @kMutagene
The new package is up. Seems like i have to be more explicit about the version? Maybe it is enough, can you try it?

![image](https://user-images.githubusercontent.com/21338071/58807519-84fcae00-8618-11e9-9d81-e56e1c55ef79.png)
