### 2019-06-03 CSBiology/FSharpGephiStreamer#15 @PrashantVaidyanathan
I tried this. I think it might be important to specify the version.
When I installed it, by default it chose Newtonsoft.Json version 3.5.8 (and also showed a warning stating that the package may not be compatible). I also ran into the same issue (during runtime) as Colin did.
However, I was able to circumvent it by installing Newstonsoft separately.

I am still unable to view the nodes. I've been getting a few runtime errors. I suspect I might be running into an issue similar to #16

OS and framework information:

OS: Windows 10
.Net core SDK version: 2.2.0

openjournals/joss-reviews#1445
