### 2019-06-04 CSBiology/FSharpGephiStreamer#16 @kMutagene
I have a hard time reproducing your error. I opened a new script and created the following dependency file in the same folder:

```
generate_load_scripts: true

source https://www.nuget.org/api/v2

nuget FSharpGephiStreamer
```

If i now create a script file with the generated loading script (`main.group.fsx`), and use the functions provided in the [streaming docs](https://csbiology.github.io/FSharpGephiStreamer/Streaming.html):

```
#load @"C:\Users\Kevin\source\scripts\GephiStreamerTest\.paket\load\netstandard2.1\main.group.fsx"

open FSharpGephiStreamer

/// Record type that represents a custom node
type MyNode = {
    Id    : int
    Label : string
    Size  : float
    Data  : string
    }

/// Record type that represents a custom edge
type MyEdge = {
    Id :     int
    Source : int
    Target : int
    Weight : float
    }

///converts a MyNode type to a list of grammar attributes
let nodeConverter (node:MyNode) =
    [
        Grammar.Attribute.Label node.Label;
        Grammar.Attribute.Size  node.Size;
        Grammar.Attribute.Color (Colors.Table.StatisticalGraphics24.getRandomColor());
        Grammar.Attribute.UserDef ("UserData",node.Data);
    ]

///converts a MyEdge type to a list of grammar attributes
let edgeConverter (edge:MyEdge) =
    [
        Grammar.Attribute.Size  edge.Weight;
        Grammar.Attribute.EdgeType  Grammar.EdgeDirection.Undirected;
        Grammar.Attribute.Color Colors.Table.Office.grey ;
    ]

///Adds a node of type MyNode using the node converter to the gephi graph
let addMyNode (node:MyNode) =
    Streamer.addNode nodeConverter node.Id node

///Adds an edge of type MyEdge using the edge converter to the gephi graph
let addMyEdge (edge: MyEdge) =
    Streamer.addEdge edgeConverter edge.Id edge.Source edge.Target edge

let myNode = {Id = 1; Label = "FOO" ; Size = 4.; Data = "BAR"}

addMyNode myNode
```

I get a node in gephi with the current nuget package:

Before (Go to the data laboratory tab in gephi):

![image](https://user-images.githubusercontent.com/21338071/58863340-f38e4a00-86b2-11e9-9354-e2ab0d33b74e.png)

After running the script:

![image](https://user-images.githubusercontent.com/21338071/58863509-3bad6c80-86b3-11e9-9e4f-37787e847a48.png)

Could you guys try this?

Edit: Regarding the issue not seeing nodes: It may also be (Most likely not though, because you guys get the HTTPCreateRequest error), that the node is just out of focus in gephi. That was exactly the case for the node generated in this script here, so you can also try to center the view on the graph:

![image](https://user-images.githubusercontent.com/21338071/58863884-fb022300-86b3-11e9-8730-af3eff375c1c.png)

which for me results in this picture:

![image](https://user-images.githubusercontent.com/21338071/58863909-07867b80-86b4-11e9-886c-1226fa5a106c.png)
