### 2019-06-04 CSBiology/FSharpGephiStreamer#16 @PrashantVaidyanathan
I'm still unable to see any nodes. When I tried the code specified above, I get the following error (I get the same error when I try the other example):
`Exception thrown: 'System.NullReferenceException' in FSharpGephiStreamer.dll`
`Exception thrown: 'System.NullReferenceException' in System.Private.CoreLib.dll`

I do see this in the console:
`{"an":{"1":{"label":"FOO","size":4.0,"color":"#8DD593","UserData":"BAR"}}}`

However I am still unable to see anything in Gephi.

I also verified that Gephi was the only process using ports 8080 and 8443.
