### 2019-06-04 CSBiology/FSharpGephiStreamer#16 @kMutagene
@cgravill :
> means that you're masking errors with the implicit ignore on the List.map (which generates warnings). I suggest you need to choose, if you're going to use the Either monad then it's important to include that in your examples

I added this to the new [basic sample](https://csbiology.github.io/FSharpGephiStreamer/SimpleNetwork.html)

Is this sufficient for you?
