### 2019-06-04 511dc64
Bump version to 1.1.2
