### 2019-06-04 CSBiology/FSharpGephiStreamer#16 @kMutagene
@PrashantVaidyanathan

> I do see this in the console:
{"an":{"1":{"label":"FOO","size":4.0,"color":"#8DD593","UserData":"BAR"}}}

This means that the Json string is created successfully. The error must be happening somewhere around here in the source code:
```

    /// Adds a node to the gephi graph using the given node converters grammar attribute mapping
    let addNode (nodeConverter:NodeConverter<'node>) nodeId =
        let nodeId' = string nodeId
        (Either.tryCatch

            // This works for you

            (fun (node:'node) ->  (jsonFormatNode Action.Add nodeId' (nodeConverter node)))
            (fun exn  -> Error.CreateJsonFailed exn.Message))
        >> Either.bind (generatePostRequest (getEnvirmonment ()) )
        >> Either.bind evalHttpResponseText
        >> Either.either
            (fun _     -> Success nodeId' )

            //This fails

            (fun error -> Failure error )
```

However i find it strange that we get a null reference exception here. I uploaded a new package with updated dependencies, I hope this can fix your error.
