### 2019-06-04 CSBiology/FSharpGephiStreamer#16 @PrashantVaidyanathan
I'm still getting a null reference exception (even with the updated dependencies). I think you are right, the error is occurring here:
`let addMyNode (node:MyNode) = `
`    Streamer.addNode nodeConverter node.Id node`

The specific error is:
`Exception thrown: 'System.NullReferenceException' in FSharpGephiStreamer.dll ("Object reference not set to an instance of an object.")`

More specifically, the nodeAddResults and edgeAddResults have the following values (for all nodes and edges):
`Http_CreateRequestFailed "Object reference not set to an instance of an object."`
