### 2019-06-04 CSBiology/FSharpGephiStreamer#16 @kMutagene
are you sure you have the right package @PrashantVaidyanathan ? The nuget page lists 0 downloads for 1.1.2:

![image](https://user-images.githubusercontent.com/21338071/58893276-1e4ac380-86f0-11e9-98b8-dcd6ce64f682.png)
