### 2019-06-04 CSBiology/FSharpGephiStreamer#16 @PrashantVaidyanathan
![Screenshot](https://user-images.githubusercontent.com/2364721/58894178-6a463a00-86e9-11e9-8770-f9141565fc0c.png)

Yes. I'm not sure why the nuget page hasn't updated yet though.
