### 2019-06-06 CSBiology/FSharpGephiStreamer#16 @kMutagene
I was finally able to reproduce the error, now that i saw that you are executing it in a console project template. When i do this i get the same:

![image](https://user-images.githubusercontent.com/21338071/59018702-c70b3700-8846-11e9-9ca9-913efebe4483.png)

While i am not sure yet why this occurs, have you executed the examples like this all the time? if so, can you try creating a `.fsx` script file and send the code directly to F# interactive? If that works for you (This is how all of this was tested), we at least enclosed the error a bit.
