### 2019-06-06 CSBiology/FSharpGephiStreamer#16 @kMutagene
I found the problem. The Http.fs version we used was encoding uri slashes with a hack using reflection and setting fields of objects. The accession of this field caused the NullReferenceException. I updated the package to use the Http.fs nuget package to handle http request creation and encoding. Streaming nodes and edges should work in program loops too.
