### 2019-06-06 CSBiology/FSharpGephiStreamer#16 @PrashantVaidyanathan
It looks like this update also requires the latest version of Hopac to be installed. I ran into an error using just version 1.1.3
However, when I installed the latest version of Hopac (https://github.com/Hopac/Hopac) via NuGet, everything started working. So perhaps you could fix this by adding Hopac as a dependency as well (version 0.4.1 or higher)?
