### 2019-06-06 CSBiology/FSharpGephiStreamer#16 @kMutagene
Yeah i just saw that, but this is a strange error as hopac is listed as dependency of Http.fs, not of the package itself. So i don't get why package management does not resolve this. I will check if adding hopac as dependency fixes the problem.
