### 2019-06-06 f0c4047
Use min strategy for transitive dependencies
