### 2019-06-06 CSBiology/FSharpGephiStreamer#16 @PrashantVaidyanathan
I can confirm that I was able to run both the examples after I installed hopac as a dependency.
