### 2019-06-06 CSBiology/FSharpGephiStreamer#16 @kMutagene
I think i found the root of this problem:

paket (the package manager we use) uses a different strategy to resolve [transitive dependencies](https://fsprojects.github.io/Paket/dependencies-file.html#Strategy-option) then nuget. This also explains the problems you guys faced with the Newtonsoft.Json package dependency:

While paket resolves transient dependencies with the maximum version, nuget takes the minimum matching version. And this is how we end up with these dependency conflicts. I was not aware of this, sorry for all the confusion.

So what i will do now is emulate the nuget manager behaviour in our paket.dependencies file so that a package can be generated that is compatible with nuget manager out of the box.

Edit:
I think i managed to make the package compatible with both paket and nuget package manager (v1.1.6)
