### 2019-06-06 3d0a577
Change paket.dependencies to nuget package manager compatible dependencies
