### 2019-06-10 CSBiology/FSharpGephiStreamer#16 @PrashantVaidyanathan
That seems to have fixed the issue. I just tested v1.1.6 and both the examples seem to work :)
