### 2019-06-10 #1445 @PrashantVaidyanathan
@danielskatz We identified some bugs Specifically https://github.com/CSBiology/FSharpGephiStreamer/issues/15 and https://github.com/CSBiology/FSharpGephiStreamer/issues/16
And it looks like these have been fixed. The authors also added an enhancement (for the documentation)

I'm close to finishing my review.
