### 2019-06-10 #1445 @cgravill
There were some issues, that sound like they've now been resolved. Some of it is on a pull request https://github.com/CSBiology/FSharpGephiStreamer/pull/23 but I'm planning to try out the test release, hopefully this week.
