### 2019-06-11 #1445 @muehlhaus
@danielskatz I think this review process is very useful and helps us a lot to improve the quality of the FSharpGephiStreamer. Thank you all for that. We resolved the issues and will close all after successful testing within the next few days.
