### 2019-06-12 CSBiology/FSharpGephiStreamer#15 @kMutagene
Closing this as this should be fixed by #23
