### 2019-06-12 #1445 @PrashantVaidyanathan
I've completed my review (and the checklist)
@danielskatz and @muehlhaus
For the point: Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?
I believe that while there isn't an explicit test in the documentation, the simple example: https://csbiology.github.io/FSharpGephiStreamer/SimpleNetwork.html serves as a manual step to ensure that the FSharpGephiStreamer works as intended.

For all other points, I think this project is well documented and the software works as intended and is very useful.
