### 2019-06-14 #1445 @cgravill
I'm satisfied with the changes that it's now clearer how to install and operate the tool.

Everything is checked off and it looks like it's ready to proceed @danielskatz
