### 2019-06-14 #1445 @muehlhaus
@danielskatz - Great. I just enabled Zenodo for our repository. The DOI is: 10.5281/zenodo.3246352
Thank you.

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3246352.svg)](https://doi.org/10.5281/zenodo.3246352)
