### 2019-06-14 #1445 @danielskatz
Is the version still 1.1.0?
