### 2019-06-14 #1445 @muehlhaus
@danielskatz The latest version is 1.1.6
