### 2019-06-14 #1445 @cgravill
Thanks for sharing your library and responding so quickly to issues. Also looking forward there being more F# in biology coming out https://github.com/CSBiology
